Neues HTML in Funktionen/Amtsbezeichnungen integriert; Default auf 10 Einträge gesetzt
git-svn-id: svn+ssh://evolvis.org/scmrepos/svn/veraweb@1043 293e2372-0500-4244-b9aa-80dbbde7b6ea

+++ trunk/veraweb-war/src/main/javaOld/de/tarent/aa/veraweb/worker/FunctionWorker.java
@@ -43,6 +43,6 @@ public class FunctionWorker extends StammdatenWorker {
     //
 	@Override
     protected Integer getLimit(OctopusContext cntx) {
-		return new Integer(5);
+		return new Integer(10);
 	}
 }
